separate opcodes implementations to directories
