+++ src/Lib/Dwarf/Expression/Opcodes/LiteralEncodings/Constant.php
@@ -11,8 +11,9 @@
 
 declare(strict_types=1);
 
-namespace PhpProfiler\Lib\Dwarf\Expression\Opcodes;
+namespace PhpProfiler\Lib\Dwarf\Expression\Opcodes\LiteralEncodings;
 
+use PhpProfiler\Lib\Dwarf\Expression\Opcodes\Opcode;
 use PhpProfiler\Lib\Dwarf\Expression\Stack;
 
 final class Constant implements Opcode
